+++ .typedoc.js
@@ -3,11 +3,15 @@ module.exports = {
   out: "./docs/",
   readme: "./README.md",
   includes: "./src",
-  gitRevision: "master", // < prevents docs from changing after every commit
+  gitRevision: "main", // < prevents docs from changing after every commit
   exclude: ["**/*test*", "**/*spec*"],
-  mode: "file",
   excludePrivate: true,
-  excludeProtected: true,
-  excludeExternals: true,
-  excludeNotExported: true
+  entryPoints: [
+    "./src/BatchCluster.ts",
+    // "./src/BatchClusterOptions.ts",
+    // "./src/BatchProcessOptions.ts",
+    // "./src/Logger.ts",
+    // "./src/Task.ts",
+  ],
+
 }